refactor(reports): Renamed models to schemas
