+++ reports/reports/schemas/_by_day.py
