+++ reports/tests/unittests/test_patch_report.py
@@ -5,7 +5,7 @@ from db_handler import DocumentNotFound
 from pymongo.errors import DuplicateKeyError, ServerSelectionTimeoutError
 
 from reports.database import Report
-from reports.models import ReportUpdate
+from reports.schemas import ReportUpdate
 from .. import utils
